using bigDecimal instead of regular primitives

+++ app/src/main/java/de/hdmstuttgart/kotlinapp/model/IAutoClicker.kt
@@ -1,11 +1,13 @@
 package de.hdmstuttgart.kotlinapp.model
 
+import java.math.BigDecimal
+
 interface IAutoClicker : Runnable {
 
-    val price : Int
+    val price : BigDecimal
     val name : String
-    val baseClicksPerSec : Double
-    var currentClicksPerSec : Double
-    fun collectClicks() : Int
+    val baseClicksPerSec : BigDecimal
+    var currentClicksPerSec : BigDecimal
+    fun collectClicks() : BigDecimal
     fun addClicker()
 }
